backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pinecone==6.0.0
python-dotenv==1.0.0
google-generativeai==0.7.0
numpy>=1.20.0
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0 